fix(xml): 'fix' crossrefs default xsds which I don't think even work?
This commit is mostly to keep track of what I needed to change in order to get crossref's XSD's to be seen as valid xsds (surprisingly hard, not sure if anyone uses these).

Basically I change the JATS xsds to have a namespace so we can refer to them in the Crossref XSDs. I'm not too familiar with XSDs so there's probably a better way to do this

One might wonder whether this is maintanable, and one might be right: I'm not sure how often CrossRefs XSDs change, but it's not never.

Another unfortunate happenstance is that this validator requires a JVM to run, which makes it harder to run in CI.